separated sciq and pubmedqa to different files
